# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  name: string;
  email?: string;
  /** @minLength 6 */
  password?: string;
  startDate?: string;
  wakeTime?: string;
  bedTime?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfileSync {
  name: string;
  wakeTime: string;
  bedTime: string;
  startDate: string;
  totalXP: number;
  highestStreak: number;
  onboardingComplete: boolean;
  activeProgramIds: string[];
}

export interface LoginResponse {
  userId: string;
  onboardingComplete: boolean;
  profile: UserProfileSync;
}

export interface MetricSync {
  id: string;
  name: string;
  category: string;
  inputType: string;
  scoreWeight: number;
  isCustom?: boolean;
  implementationIntention?: string | null;
}

export interface RegisterResponse {
  userId: string;
  profile: UserProfileSync;
  metrics: MetricSync[];
}

export interface DailyLogSync {
  id: string;
  metricId: string;
  date: string;
  value: number;
  note?: string | null;
}

export interface JournalEntrySync {
  id: string;
  date: string;
  prompt: string;
  response: string;
  mood: number;
  energy: number;
  tags?: string[];
  wordCount?: number;
}

export interface RelapseLogSync {
  id: string;
  date: string;
  metricId: string;
  triggerCategory: string;
  triggerReflection: string;
  nextAction: string;
  compassionStatement?: string | null;
}

export interface ProgramProgressSync {
  programId: string;
  currentWeek: number;
  weekStartDate: string;
  completedWeeks: number[];
  resetCount: number;
}

export interface WeekTaskProgressSync {
  programId: string;
  weekNumber: number;
  taskId: string;
  completed: boolean;
}

export interface FocusLogSync {
  id: string;
  date: string;
  minutes: number;
}

export interface UserDataResponse {
  profile: UserProfileSync;
  metrics: MetricSync[];
  dailyLogs: DailyLogSync[];
  journalEntries: JournalEntrySync[];
  relapseLogs: RelapseLogSync[];
  programProgress: ProgramProgressSync[];
  weekTaskProgress: WeekTaskProgressSync[];
  focusLogs: FocusLogSync[];
}

export interface SyncRequest {
  profile?: UserProfileSync;
  metrics?: MetricSync[];
  dailyLogs?: DailyLogSync[];
  journalEntries?: JournalEntrySync[];
  relapseLogs?: RelapseLogSync[];
  programProgress?: ProgramProgressSync[];
  weekTaskProgress?: WeekTaskProgressSync[];
  focusLogs?: FocusLogSync[];
}

export interface SyncResponse {
  success: boolean;
  syncedAt: string;
}

export interface PushTokenRequest {
  token: string;
}