# lib/api-zod/src/generated/types/journalEntrySync.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntrySync {
  id: string;
  date: string;
  prompt: string;
  response: string;
  mood: number;
  energy: number;
  tags?: string[];
  wordCount?: number;
}