# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileSync } from './userProfileSync';

export interface LoginResponse {
  userId: string;
  onboardingComplete: boolean;
  profile: UserProfileSync;
}