# lib/api-zod/src/generated/types/userProfileSync.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileSync {
  name: string;
  wakeTime: string;
  bedTime: string;
  startDate: string;
  totalXP: number;
  highestStreak: number;
  onboardingComplete: boolean;
  activeProgramIds: string[];
}